add app debug permission check

+++ device/services/ipc/src/unix_socket_server.cpp
@@ -22,6 +22,7 @@
 #include <unistd.h>
 #include <linux/un.h>
 
+#include "init_socket.h"
 #include "logging.h"
 #include "securec.h"
 
@@ -103,25 +104,30 @@ const int UNIX_SOCKET_LISTEN_COUNT = 5;
 bool UnixSocketServer::StartServer(const std::string& addrname, ServiceEntry& p)
 {
     CHECK_TRUE(socketHandle_ == -1, false, "StartServer FAIL socketHandle_ != -1");
+    int sock = -1;
+    if (getuid() == 0) {
+        struct sockaddr_un addr;
+        sock = socket(AF_UNIX, SOCK_STREAM, 0);
+        CHECK_TRUE(sock != -1, false, "StartServer FAIL create socket ERR : %d", errno);
 
-    struct sockaddr_un addr;
-    int sock = socket(AF_UNIX, SOCK_STREAM, 0);
-    CHECK_TRUE(sock != -1, false, "StartServer FAIL create socket ERR : %d", errno);
+        if (memset_s(&addr, sizeof(struct sockaddr_un), 0, sizeof(struct sockaddr_un)) != EOK) {
+            HILOG_ERROR(LOG_CORE, "memset_s error!");
+        }
+        addr.sun_family = AF_UNIX;
+        if (strncpy_s(addr.sun_path, sizeof(addr.sun_path), addrname.c_str(), sizeof(addr.sun_path) - 1) != EOK) {
+            HILOG_ERROR(LOG_CORE, "strncpy_s error!");
+        }
+        unlink(addrname.c_str());
+        CHECK_TRUE(bind(sock, (struct sockaddr*)&addr, sizeof(struct sockaddr_un)) == 0, close(sock) != 0,
+                   "StartServer FAIL bind ERR : %d", errno);
 
-    if (memset_s(&addr, sizeof(struct sockaddr_un), 0, sizeof(struct sockaddr_un)) != EOK) {
-        HILOG_ERROR(LOG_CORE, "memset_s error!");
+        std::string chmodCmd = "chmod 666 " + addrname;
+        HILOG_INFO(LOG_CORE, "chmod command : %s", chmodCmd.c_str());
+        std::unique_ptr<FILE, decltype(&pclose)> pipe(popen(chmodCmd.c_str(), "r"), pclose);
+    } else {
+        sock = GetControlSocket(addrname.c_str());
+        CHECK_TRUE(sock != -1, false, "StartServer FAIL GetControlSocket retrun : %d", sock);
     }
-    addr.sun_family = AF_UNIX;
-    if (strncpy_s(addr.sun_path, sizeof(addr.sun_path), addrname.c_str(), sizeof(addr.sun_path) - 1) != EOK) {
-        HILOG_ERROR(LOG_CORE, "strncpy_s error!");
-    }
-    unlink(addrname.c_str());
-    CHECK_TRUE(bind(sock, (struct sockaddr*)&addr, sizeof(struct sockaddr_un)) == 0, close(sock) != 0,
-               "StartServer FAIL bind ERR : %d", errno);
-
-    std::string chmodCmd = "chmod 666 " + addrname;
-    HILOG_INFO(LOG_CORE, "chmod command : %s", chmodCmd.c_str());
-    std::unique_ptr<FILE, decltype(&pclose)> pipe(popen(chmodCmd.c_str(), "r"), pclose);
 
     CHECK_TRUE(listen(sock, UNIX_SOCKET_LISTEN_COUNT) != -1, close(sock) != 0 && unlink(addrname.c_str()) == 0,
                "StartServer FAIL listen ERR : %d", errno);
